Move `PostDashboard` to containers, fix syntax errors
